clean and fix problem with patchin logreg
