+++ daal4py/sklearn/linear_model/__init__.py
@@ -16,8 +16,7 @@
 #===============================================================================
 
 from .linear import LinearRegression
-from .logistic_path import (logistic_regression_path,
-                            LogisticRegression)
+from .logistic_path import logistic_regression_path, LogisticRegression
 from .ridge import Ridge
 from .coordinate_descent import ElasticNet, Lasso
